requirements.txt
Flask
Flask-PyMongo
Flask-WTF
Flask-Login
pymongo
numpy
pandas
opencv-python
Pillow
scikit-learn
tensorflow
requests
python-dotenv
gunicorn
